+++ servant-client/test/Servant/ClientSpec.hs
@@ -45,7 +45,7 @@ import           Servant.API
 import           Servant.API.Authentication
 import           Servant.Client
 import qualified Servant.Common.Req         as SCR
-import           Servant.Client.Authentication (AuthenticateRequest(authReq))
+import           Servant.Client.Authentication()
 import           Servant.Server
 import           Servant.Server.Internal.Authentication
 
@@ -112,11 +112,6 @@ basicAuthCheck :: BasicAuth "realm" -> IO (Maybe Person)
 basicAuthCheck (BasicAuth user pass) = if user == "servant" && pass == "server"
                                        then return (Just $ Person "servant" 17)
                                        else return Nothing
-
-instance AuthenticateRequest (BasicAuth realm) where
-    authReq (BasicAuth user pass) req =
-        let authText = TE.decodeUtf8 ("Basic " <> B64.encode (user <> ":" <> pass)) in
-            SCR.addHeader "Authorization" authText req
 
 api :: Proxy Api
 api = Proxy
@@ -322,7 +317,7 @@ failSpec = withFailServer $ \ baseUrl -> do
        :<|> _ )
          = client api baseUrl
       getGetWrongHost :: EitherT ServantError IO Person
-      (getGetWrongHost :<|> _) = client api (BaseUrl Http "127.0.0.1" 19872)
+      (getGetWrongHost :<|> _) = client api (BaseUrl Http "127.0.0.1" 19872 "")
 
   hspec $ do
     context "client returns errors appropriately" $ do
@@ -383,7 +378,7 @@ withWaiDaemon mkApplication action = do
       runSettingsSocket settings socket application)
             `finally` notifyKilled ()
     krakenPort <- waitForStart
-    let baseUrl = (BaseUrl Http "localhost" 80){baseUrlPort = krakenPort}
+    let baseUrl = (BaseUrl Http "localhost" 80 ""){baseUrlPort = krakenPort}
     return (thread, waitForKilled, baseUrl)
   free (thread, waitForKilled, _) = do
     killThread thread