+++ servant-docs/src/Servant/Docs.hs
@@ -147,7 +147,8 @@ module Servant.Docs
   , ExtraInfo(..), docsWith, docsWithIntros, extraInfo
 
   , -- * Classes you need to implement for your types
-    ToSample(..)
+    ToAuthInfo(..)
+  , ToSample(..)
   , sampleByteString
   , sampleByteStrings
   , ToParam(..)
@@ -157,12 +158,13 @@ module Servant.Docs
     Method(..)
   , Endpoint, path, method, defEndpoint
   , API, apiIntros, apiEndpoints, emptyAPI
+  , AuthenticationInfo(..), authIntro, authDataRequired
   , DocCapture(..), capSymbol, capDesc
   , DocQueryParam(..), ParamKind(..), paramName, paramValues, paramDesc, paramKind
   , DocNote(..), noteTitle, noteBody
   , DocIntro(..), introTitle, introBody
   , Response(..), respStatus, respTypes, respBody, defResponse
-  , Action, captures, headers, notes, params, rqtypes, rqbody, response, defAction
+  , Action, authInfo, captures, headers, notes, params, rqtypes, rqbody, response, defAction
   , single
   ) where